@@ -1,59 +1,62 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-  </PropertyGroup>
+    <PropertyGroup>
+        <TargetFramework>net5.0</TargetFramework>
+    </PropertyGroup>
 
-  <ItemGroup>
-    <None Remove="Scripts\controls-user.js" />
-  </ItemGroup>
+    <ItemGroup>
+        <None Remove="Scripts\controls-user.js"/>
+    </ItemGroup>
 
-  <ItemGroup>
-    <Content Include="Scripts\controls\controls-user.js" />
-  </ItemGroup>
+    <ItemGroup>
+        <Content Include="Scripts\controls\controls-user.js"/>
+    </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="jQuery" Version="3.4.1" />
-    <PackageReference Include="Microsoft.AspNet.Identity.Core" Version="2.2.2" />
-    <PackageReference Include="Microsoft.AspNet.Identity.Owin" Version="2.2.2" />
-    <PackageReference Include="Microsoft.AspNet.WebApi.Owin" Version="5.2.7" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.1.2" PrivateAssets="All" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Language" Version="2.2.0" />
-    <PackageReference Include="Microsoft.Owin.Host.SystemWeb" Version="4.0.1" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="2.1.9" />
-    <PackageReference Include="System.Net.Http" Version="4.3.4" />
-  </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="jQuery" Version="3.6.0"/>
+        <PackageReference Include="Microsoft.AspNet.Identity.Core" Version="2.2.3"/>
+        <PackageReference Include="Microsoft.AspNet.Identity.Owin" Version="2.2.3"/>
+        <PackageReference Include="Microsoft.AspNet.WebApi.Owin" Version="5.2.7"/>
+        <PackageReference Include="Microsoft.AspNetCore.App"/>
+        <PackageReference Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="5.0.5"/>
+        <PackageReference Include="Microsoft.Owin.Host.SystemWeb" Version="4.1.1"/>
+        <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="5.0.2"/>
+        <PackageReference Include="System.Net.Http" Version="4.3.4"/>
+    </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Scripts\Libraries\" />
-  </ItemGroup>
+    <ItemGroup>
+        <Folder Include="Scripts\Libraries\"/>
+    </ItemGroup>
 
-  <ItemGroup>
-    <None Include="wwwroot\js\Scripts\shared.js" />
-    <None Include="wwwroot\lib\axios\axios.min.js">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="wwwroot\lib\polyfill\polyfill.min.js">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="wwwroot\lib\vue\vue.js">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Include="wwwroot\lib\vue\vue.min.js">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
+    <ItemGroup>
+        <None Include="wwwroot\js\Scripts\shared.js"/>
+        <None Include="wwwroot\lib\axios\axios.min.js">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+        <None Include="wwwroot\lib\polyfill\polyfill.min.js">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+        <None Include="wwwroot\lib\vue\vue.js">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+        <None Include="wwwroot\lib\vue\vue.min.js">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </None>
+    </ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\MvcFrameworkBll\MvcFrameworkBll.csproj" />
-    <ProjectReference Include="..\MvcFrameworkCml\MvcFrameworkCml.csproj" />
-  </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\MvcFrameworkBll\MvcFrameworkBll.csproj"/>
+        <ProjectReference Include="..\MvcFrameworkCml\MvcFrameworkCml.csproj"/>
+    </ItemGroup>
 
-  <ItemGroup>
-    <Content Update="appsettings.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </Content>
-  </ItemGroup>
+    <ItemGroup>
+        <Content Update="appsettings.json">
+            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+        </Content>
+    </ItemGroup>
+
+    <PropertyGroup>
+        <AddRazorSupportForMvc>true</AddRazorSupportForMvc>
+    </PropertyGroup>
 
 </Project>
