@@ -13,7 +13,6 @@ namespace MvcFrameworkWeb.Controllers
 {
     public class CustomBaseController : Controller
     {
-
         protected readonly ControllerHelper _controllerHelper_;
         protected readonly IAppSettings _appSettings_;
         protected readonly ILogger _logger_;
@@ -28,15 +27,22 @@ namespace MvcFrameworkWeb.Controllers
         protected IActionResult GetViewModelOrRedirect(HttpContext httpContext)
         {
             if (httpContext.User.Identity.IsAuthenticated)
-                return View(new BaseViewModel() { User = new EndUser(httpContext.User) });
+                return View(new BaseViewModel()
+                {
+                    User = new EndUser(httpContext.User)
+                });
             else
             {
-                if (httpContext.Request.Path.HasValue &&
-                    httpContext.Request.Path.Value.ToLower().Contains("login"))
-                    return View();
-                if (httpContext.Request.Path.HasValue &&
-                    httpContext.Request.Path.Value.ToLower().Contains("register"))
-                    return View("~/Views/User/Register.cshtml");
+                if (httpContext.Request.Path.Value != null)
+                {
+                    if (httpContext.Request.Path.HasValue &&
+                        httpContext.Request.Path.Value.ToLower().Contains("login"))
+                        return View();
+                    if (httpContext.Request.Path.HasValue &&
+                        httpContext.Request.Path.Value.ToLower().Contains("register"))
+                        return View("~/Views/User/Register.cshtml");
+                }
+
                 var action = GetRedirectAction(httpContext);
                 return RedirectToAction(action.Item1, action.Item2);
             }
@@ -49,11 +55,13 @@ namespace MvcFrameworkWeb.Controllers
             {
                 return ("Register", "User");
             }
+
             var idClaimValue = httpContext.User.Claims.FirstOrDefault(x => x.Type.Equals(ClaimTypesExt.Id))?.Value;
             if (Int32.TryParse(idClaimValue, out Int32 id) && id > 0)
             {
                 return ("Index", "Home");
             }
+
             return ("Index", "Login");
         }
     }
